Split OCI dir I/O out of legacy.rs and clarify what "legacy" means
The previous `legacy.rs` mixed two unrelated concerns under a single
"legacy" label:

1. Generic OCI Image Layout directory I/O (read `index.json`, parse
   `ImageManifest`, copy CAS blobs into the v3 registry).
2. v2 OMMX local registry layout helpers (recursive scan of a
   `<root>/<image_name>/<tag>/` tree, path-to-`ImageName` resolution,
   batch import with per-directory aggregation).

(1) is **not legacy**: an OCI Image Layout directory is the OCI
standard, produced by `oras` / `crane` / `skopeo` and used by v3
itself as the import / export interchange. (2) **is** legacy: the
"path/tag tree of OCI dirs" layout was OMMX v2's local registry shape
and v3 explicitly replaces it with the SQLite-backed registry.

Now reflected in the code:

- New module `oci_dir.rs` owns generic OCI Image Layout I/O —
  `import_oci_dir`, `import_oci_dir_with_policy`,
  `import_oci_dir_as_ref(_with_policy)`, `oci_dir_ref`,
  `oci_dir_image_name`, the `OciDirRef` identity tuple, the
  `RefConflictHandling` enum, and the `prepare_oci_dir` /
  `read_oci_dir_blob` / `oci_dir_blob_path` helpers. Module-level
  doc-comment makes the "not legacy" intent explicit.
- `legacy.rs` shrinks to v2-specific code only — `legacy_local_registry_path`,
  `gather_legacy_oci_dirs`, `legacy_import_image_name`,
  `import_legacy_local_registry(_ref)(_with_policy)` and the
  `LegacyImportReport` aggregator. It uses `oci_dir.rs` internally.
- `local_registry/mod.rs` re-exports both modules with a docstring
  that lays out the layering (`index` / `blob` / `types` →
  `registry` → `oci_dir` → `legacy`).
- Function and struct rename consequences from the previous "single
  OCI dir is not legacy" rename pass are now reflected in the file
  layout, not just identifier prefixes.

Validation: cargo test -p ommx --lib (501 passed, 16 in artifact::*),
cargo clippy -p ommx --lib unchanged (3 pre-existing warnings),
cargo fmt clean.
